qoder-enterprise-system-py/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "qoder-enterprise-py"
version = "1.0.0"
description = "Python microservices for Qoder Enterprise System"
authors = ["CTO Architecture Team <arch-team@company.com>"]
readme = "README.md"
packages = [
    { include = "qoder_enterprise", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"

# Web Framework
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.32.0" }

# Workflow & Task Queue
celery = { extras = ["redis"], version = "^5.4.0" }
redis = "^5.2.0"

# Data Processing
pydantic = "^2.9.0"
pydantic-settings = "^2.6.0"
pyyaml = "^6.0.2"

# NLP & ML
transformers = "^4.46.0"
torch = "^2.5.0"
sentence-transformers = "^3.2.0"
scikit-learn = "^1.5.0"
spacy = "^3.8.0"

# Git Operations
gitpython = "^3.1.43"

# Utilities
httpx = "^0.27.0"
structlog = "^24.4.0"
python-json-logger = "^2.0.7"
tenacity = "^9.0.0"

# Testing
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
factory-boy = "^3.3.0"

# Dev Tools
black = "^24.10.0"
ruff = "^0.7.0"
mypy = "^1.13.0"
pre-commit = "^4.0.0"

[tool.poetry.scripts]
qoder-hooks = "qoder_enterprise.hooks.cli:main"
qoder-nlp = "qoder_enterprise.nlp.cli:main"

[tool.poetry.group.dev.dependencies]
ipython = "^8.29.0"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "W",  # pydocstyle
    "UP", # pyupgrade
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = ["E501"]  # Line too long (handled by black)

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
]

[tool.coverage.run]
source = ["src/qoder_enterprise"]
omit = ["*/tests/*", "*/test_*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
